# app/api/routes/media.py
from fastapi import APIRouter, UploadFile, File, Form, Depends, HTTPException
from typing import Optional
from app.schemas.media import MediaAssetResponse
import uuid

router = APIRouter()

@router.post("/upload", response_model=MediaAssetResponse, status_code=201)
async def upload_media(
    file: UploadFile = File(...),
    asset_type: str = Form(...),
    reference_id: Optional[str] = Form(None)
):
    return {
        "id": uuid.uuid4(),
        "user_id": uuid.uuid4(),
        "original_name": file.filename,
        "file_size_bytes": 1024,
        "mime_type": file.content_type,
        "asset_type": asset_type,
        "reference_id": reference_id,
        "s3_key": f"{asset_type}/{uuid.uuid4()}-{file.filename}",
        "uploaded_at": "2026-05-31T12:00:00Z"
    }

@router.get("/{media_id}", response_model=MediaAssetResponse)
async def get_media_metadata(media_id: uuid.UUID):
    raise HTTPException(status_code=404, detail="Not implemented")